Fix crit chances for normal attack
The critical modifier `crit_mod` was being divided by 10 twice for normal attacks but only once for magic and talk. This would make it require at least 100 luck or dexterity for it to affect critical chance. 
Empirical evidence seems to support this, as normal attacks frequently fumble even with high luck, but magic attacks of most players seem never to.